fix(wav): write bits_per_sample as u16 in streaming writer fmt chunks
ValidatedSampleType::bits_per_sample() returns a NonZeroUsize, so
calling to_le_bytes() on .get() serialized 8 bytes where the fmt chunk
field is 2 bytes wide. Every WAV produced by StreamedWavWriter carried
6 stray zero bytes after the declared fmt chunk, so readers found a
zeroed chunk id where the data chunk belonged and rejected the file.

+++ src/wav/streaming_writer.rs
@@ -198,7 +198,7 @@ where
         writer.write_all(&sample_rate.to_le_bytes())?;
         writer.write_all(&byte_rate.to_le_bytes())?;
         writer.write_all(&block_align.to_le_bytes())?;
-        writer.write_all(&bits_per_sample.get().to_le_bytes())?;
+        writer.write_all(&(bits_per_sample.get() as u16).to_le_bytes())?;
 
         Ok(())
     }
@@ -245,13 +245,13 @@ where
         writer.write_all(&sample_rate.to_le_bytes())?;
         writer.write_all(&byte_rate.to_le_bytes())?;
         writer.write_all(&block_align.to_le_bytes())?;
-        writer.write_all(&bits_per_sample.get().to_le_bytes())?;
+        writer.write_all(&(bits_per_sample.get() as u16).to_le_bytes())?;
 
         // Extension size (2 bytes)
         writer.write_all(&22u16.to_le_bytes())?;
 
         // Extended format info (22 bytes)
-        writer.write_all(&bits_per_sample.get().to_le_bytes())?; // Valid bits per sample
+        writer.write_all(&(bits_per_sample.get() as u16).to_le_bytes())?; // Valid bits per sample
         writer.write_all(&channel_mask.to_le_bytes())?;
 
         // Sub-format GUID (16 bytes)
